move StartContext to server
