+++ packages/start/entry-client/StartClient.tsx
@@ -1,6 +1,6 @@
 import { MetaProvider } from "solid-meta";
 import { Router } from "solid-app-router";
-import { StartProvider } from "../entry-server/StartContext";
+import { StartProvider } from "../server/StartContext";
 import Root from "~/root";
 
 const rootData = Object.values(import.meta.globEager("/src/root.data.(js|ts)"))[0];
