+++ packages/start/entry-server/index.ts
@@ -1,8 +1,7 @@
-
-export * from "./StartContext";
-export * from "./entries";
+export * from "../server/StartContext";
+export * from "./render";
 
 // server-side only exports
 export { default as StartServer } from "./StartServer";
 export { createHandler, composeMiddleware } from "./StartServer";
-export type { RequestContext, Middleware, MiddlewareInput, MiddlewareFn } from "./StartServer";
+export type { Middleware, MiddlewareInput, MiddlewareFn } from "./StartServer";
