+++ packages/start/root/Links.tsx
@@ -1,7 +1,7 @@
 import { useContext, JSXElement } from "solid-js";
 import { Assets } from "solid-js/web";
-import { StartContext } from "../entry-server/StartContext";
-import { ContextMatches, ManifestEntry, RequestContext } from "../entry-server/StartServer";
+import { StartContext } from "../server/StartContext";
+import { ContextMatches, ManifestEntry, RequestContext } from "../server/types";
 
 function getAssetsFromManifest(
   manifest: RequestContext["manifest"],
